+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[FIX]** Fix pointy corner on active `Message` (from top-right to bottom-right)
 - **[BREAKING CHANGE]** `<Text>`: changed export for `TextDisplayType` `TextTagType`
 - **[BREAKING CHANGE]** `<Tabs>`: changed export for `TabsStatus`
+- **[BREAKING CHANGE]** `Button`: changed export for `Button.STATUS` to new `ButtonStatus`
 
 # v8.8.0 (24/07/2019)